Per comments moved StartJobSerializer to utils and added tests
