+++ job-server/src/main/resources/application.conf
@@ -165,7 +165,7 @@ akka {
     provider = "akka.cluster.ClusterActorRefProvider"
     warn-about-java-serializer-usage = off
     serializers {
-      customStartJobSerializer = "spark.jobserver.JobManagerActor$StartJobSerializer"
+      customStartJobSerializer = "spark.jobserver.util.StartJobSerializer"
     }
     serialization-bindings {
       "spark.jobserver.JobManagerActor$StartJob" = customStartJobSerializer
