+++ job-server/src/main/scala/spark/jobserver/JobManagerActor.scala
@@ -3,11 +3,9 @@ package spark.jobserver
 import java.net.{URI, URL}
 import java.util.concurrent.Executors._
 import java.util.concurrent.atomic.AtomicInteger
-import java.io.{ByteArrayInputStream, ByteArrayOutputStream, ObjectInputStream, ObjectOutputStream}
 
 import akka.actor.{ActorRef, PoisonPill, Props}
-import akka.serialization.JSerializer
-import com.typesafe.config.{Config, ConfigFactory, ConfigRenderOptions}
+import com.typesafe.config.Config
 import org.apache.hadoop.conf.Configuration
 import org.apache.spark.{SparkConf, SparkEnv}
 import org.joda.time.DateTime
@@ -20,8 +18,6 @@ import spark.jobserver.util.{ContextURLClassLoader, SparkJobUtils}
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success, Try}
 import spark.jobserver.common.akka.InstrumentedActor
-
-import org.slf4j.LoggerFactory
 
 object JobManagerActor {
   // Messages
@@ -45,56 +41,6 @@ object JobManagerActor {
   // Akka 2.2.x style actor props for actor creation
   def props(contextConfig: Config, daoActor: ActorRef): Props = Props(classOf[JobManagerActor],
     contextConfig, daoActor)
-
-  // Serializer StartJob message
-  class StartJobSerializer extends JSerializer {
-    val logger = LoggerFactory.getLogger(getClass)
-
-    override def includeManifest() : Boolean = false
-
-    override def identifier() : Int = 12376
-
-    override def toBinary(obj: AnyRef): Array[Byte] = {
-      logger.debug(s"Serializing StartJob object -- ${obj}")
-      try {
-        val byteArray = new ByteArrayOutputStream()
-        val out = new ObjectOutputStream(byteArray)
-        val startJob = obj.asInstanceOf[StartJob]
-        out.writeObject(startJob.appName)
-        out.writeObject(startJob.classPath)
-        out.writeObject(startJob.config.root().render(ConfigRenderOptions.concise()))
-        out.writeObject(startJob.subscribedEvents)
-        out.flush()
-        byteArray.toByteArray
-      } catch {
-        case ex : Exception =>
-          throw new IllegalArgumentException(s"Object of unknown class cannot be serialized " +
-            s"${ex.getCause.getMessage}")
-      }
-    }
-
-    override def fromBinaryJava(bytes: Array[Byte],
-                                clazz: Class[_]): AnyRef = {
-      logger.debug(s"Deserializing StartJob object -- ${bytes.length}")
-      try {
-        val input = new ByteArrayInputStream(bytes)
-        val inputStream = new ObjectInputStream(input)
-        val appName = inputStream.readObject().asInstanceOf[String]
-        val classPath = inputStream.readObject().asInstanceOf[String]
-        val configString = inputStream.readObject().asInstanceOf[String]
-        val subscribedEvents = inputStream.readObject().asInstanceOf[Set[Class[_]]]
-        logger.debug(s"appname: ${appName}")
-        logger.debug(s"classPath: ${classPath}")
-        logger.debug(s"configString: ${configString}")
-        logger.debug(s"subscribedEvents: ${subscribedEvents}")
-        StartJob(appName, classPath, ConfigFactory.parseString(configString), subscribedEvents)
-      } catch {
-        case ex: Exception =>
-          throw new IllegalArgumentException(s"Object of unknown class cannot be deserialized " +
-            s"${ex.getCause.getMessage}")
-      }
-    }
-  }
 }
 
 /**
