+++ job-server/src/test/scala/spark/jobserver/JobManagerActorSpec.scala
@@ -49,5 +49,4 @@ class JobManagerActorSpec extends JobManagerSpec {
       }
     }
   }
-
 }
